Module aws_sdk_ecrpublic::error

Modules

empty_upload_exception

See EmptyUploadException

image_already_exists_exception

See ImageAlreadyExistsException

image_digest_does_not_match_exception

See ImageDigestDoesNotMatchException

image_not_found_exception

See ImageNotFoundException

image_tag_already_exists_exception

See ImageTagAlreadyExistsException

invalid_layer_exception

See InvalidLayerException

invalid_layer_part_exception

See InvalidLayerPartException

invalid_parameter_exception

See InvalidParameterException

invalid_tag_parameter_exception

See InvalidTagParameterException

layer_already_exists_exception

See LayerAlreadyExistsException

layer_part_too_small_exception

See LayerPartTooSmallException

layers_not_found_exception

See LayersNotFoundException

limit_exceeded_exception

See LimitExceededException

referenced_images_not_found_exception

See ReferencedImagesNotFoundException

registry_not_found_exception

See RegistryNotFoundException

repository_already_exists_exception

See RepositoryAlreadyExistsException

repository_not_empty_exception

See RepositoryNotEmptyException

repository_not_found_exception

See RepositoryNotFoundException

repository_policy_not_found_exception

See RepositoryPolicyNotFoundException

server_exception

See ServerException

too_many_tags_exception

See TooManyTagsException

unsupported_command_exception

See UnsupportedCommandException

upload_not_found_exception

See UploadNotFoundException

Structs

BatchCheckLayerAvailabilityError

BatchDeleteImageError

CompleteLayerUploadError

CreateRepositoryError

DeleteRepositoryError

DeleteRepositoryPolicyError

DescribeImageTagsError

DescribeImagesError

DescribeRegistriesError

DescribeRepositoriesError

EmptyUploadException

The specified layer upload does not contain any layer parts.

GetAuthorizationTokenError

GetRegistryCatalogDataError

GetRepositoryCatalogDataError

GetRepositoryPolicyError

ImageAlreadyExistsException

The specified image has already been pushed, and there were no changes to the manifest

ImageDigestDoesNotMatchException

The specified image digest does not match the digest that Amazon ECR calculated for the

ImageNotFoundException

The image requested does not exist in the specified repository.

ImageTagAlreadyExistsException

The specified image is tagged with a tag that already exists. The repository is

InitiateLayerUploadError

InvalidLayerException

The layer digest calculation performed by Amazon ECR upon receipt of the image layer does not

InvalidLayerPartException

The layer part size is not valid, or the first byte specified is not consecutive to the

InvalidParameterException

The specified parameter is invalid. Review the available parameters for the API

InvalidTagParameterException

An invalid parameter has been specified. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

LayerAlreadyExistsException

The image layer already exists in the associated repository.

LayerPartTooSmallException

Layer parts must be at least 5 MiB in size.

LayersNotFoundException

The specified layers could not be found, or the specified layer is not valid for this

LimitExceededException

The operation did not succeed because it would have exceeded a service limit for your

ListTagsForResourceError

PutImageError

PutRegistryCatalogDataError

PutRepositoryCatalogDataError

ReferencedImagesNotFoundException

The manifest list is referencing an image that does not exist.

RegistryNotFoundException

The registry does not exist.

RepositoryAlreadyExistsException

The specified repository already exists in the specified registry.

RepositoryNotEmptyException

The specified repository contains images. To delete a repository that contains images,

RepositoryNotFoundException

The specified repository could not be found. Check the spelling of the specified

RepositoryPolicyNotFoundException

The specified repository and registry combination does not have an associated repository

ServerException

These errors are usually caused by a server-side issue.

SetRepositoryPolicyError

TagResourceError

TooManyTagsException

The list of tags on the repository is over the limit. The maximum number of tags that

UnsupportedCommandException

The action is not supported in this Region.

UntagResourceError

UploadLayerPartError

UploadNotFoundException

The upload could not be found, or the specified upload ID is not valid for this

Enums

BatchCheckLayerAvailabilityErrorKind

BatchDeleteImageErrorKind

CompleteLayerUploadErrorKind

CreateRepositoryErrorKind

DeleteRepositoryErrorKind

DeleteRepositoryPolicyErrorKind

DescribeImageTagsErrorKind

DescribeImagesErrorKind

DescribeRegistriesErrorKind

DescribeRepositoriesErrorKind

GetAuthorizationTokenErrorKind

GetRegistryCatalogDataErrorKind

GetRepositoryCatalogDataErrorKind

GetRepositoryPolicyErrorKind

InitiateLayerUploadErrorKind

ListTagsForResourceErrorKind

PutImageErrorKind

PutRegistryCatalogDataErrorKind

PutRepositoryCatalogDataErrorKind

SetRepositoryPolicyErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UploadLayerPartErrorKind