Module aws_sdk_ecrpublic::model

Modules

authorization_data

See AuthorizationData

image

See Image

image_detail

See ImageDetail

image_failure

See ImageFailure

image_identifier

See ImageIdentifier

image_tag_detail

See ImageTagDetail

layer

See Layer

layer_failure

See LayerFailure

referenced_image_detail

See ReferencedImageDetail

registry

See Registry

registry_alias

See RegistryAlias

registry_catalog_data

See RegistryCatalogData

repository

See Repository

repository_catalog_data

See RepositoryCatalogData

repository_catalog_data_input

See RepositoryCatalogDataInput

tag

See Tag

Structs

AuthorizationData

An authorization token data object that corresponds to a public registry.

Image

An object representing an Amazon ECR image.

ImageDetail

An object that describes an image returned by a DescribeImages

ImageFailure

An object representing an Amazon ECR image failure.

ImageIdentifier

An object with identifying information for an Amazon ECR image.

ImageTagDetail

An object representing the image tag details for an image.

Layer

An object representing an Amazon ECR image layer.

LayerFailure

An object representing an Amazon ECR image layer failure.

ReferencedImageDetail

An object that describes the image tag details returned by a DescribeImageTags action.

Registry

The details of a public registry.

RegistryAlias

An object representing the aliases for a public registry. A public registry is given an

RegistryCatalogData

The metadata for a public registry.

Repository

An object representing a repository.

RepositoryCatalogData

The catalog data for a repository. This data is publicly visible in the

RepositoryCatalogDataInput

An object containing the catalog data for a repository. This data is publicly visible in

Tag

The metadata that you apply to a resource to help you categorize and organize them. Each

Enums

ImageFailureCode

LayerAvailability

LayerFailureCode

RegistryAliasStatus