Module aws_sdk_ecrpublic::operation

Structs

BatchCheckLayerAvailability

Checks the availability of one or more image layers within a repository in a public

BatchDeleteImage

Deletes a list of specified images within a repository in a public registry. Images are

CompleteLayerUpload

Informs Amazon ECR that the image layer upload has completed for a specified public registry,

CreateRepository

Creates a repository in a public registry. For more information, see Amazon ECR

DeleteRepository

Deletes a repository in a public registry. If the repository contains images, you must

DeleteRepositoryPolicy

Deletes the repository policy associated with the specified repository.

DescribeImageTags

Returns the image tag details for a repository in a public registry.

DescribeImages

Returns metadata about the images in a repository in a public registry.

DescribeRegistries

Returns details for a public registry.

DescribeRepositories

Describes repositories in a public registry.

GetAuthorizationToken

Retrieves an authorization token. An authorization token represents your IAM

GetRegistryCatalogData

Retrieves catalog metadata for a public registry.

GetRepositoryCatalogData

Retrieve catalog metadata for a repository in a public registry. This metadata is

GetRepositoryPolicy

Retrieves the repository policy for the specified repository.

InitiateLayerUpload

Notifies Amazon ECR that you intend to upload an image layer.

ListTagsForResource

List the tags for an Amazon ECR Public resource.

PutImage

Creates or updates the image manifest and tags associated with an image.

PutRegistryCatalogData

Create or updates the catalog data for a public registry.

PutRepositoryCatalogData

Creates or updates the catalog data for a repository in a public registry.

SetRepositoryPolicy

Applies a repository policy to the specified public repository to control access

TagResource

Associates the specified tags to a resource with the specified resourceArn.

UntagResource

Deletes specified tags from a resource.

UploadLayerPart

Uploads an image layer part to Amazon ECR.